果物マスタ管理 削除機能をViewで実装し直し

+++ sales_management_system/views.py
@@ -4,7 +4,7 @@ from django.shortcuts import redirect
 from django.views import View
 
 from .forms import FruitForm
-from .models import Fruit
+from .models import Fruit, Sale
 
 
 class FruitListView(ListView):
@@ -32,11 +32,10 @@ class FruitUpdateView(UpdateView):
 class FruitDeleteView(View):
     def get(self, request, pk):
 
-        try:
-            fruit = Fruit.objects.get(id=pk)
-        except Fruit.DoesNotExist:
-            return redirect('fruit_list')
-        else:
+        fruit = Fruit.objects.get(id=pk)
+        sales_count = Sale.objects.filter(fruit_name=fruit.name).count()
+
+        if sales_count != 0:
             fruit.delete()
 
         return redirect('fruit_list')
